Prediction Algo Working for all S&P
Prediction algorithm is working, just need to add some sleep statements so that it doesn't stop early. Also, if reading from wikipedia, need to add a line that replaces BF.B with BF-B. Also also need to make this a python file. Also also also need to move this to the GPU server.